separate configure flags from code, be less intrusive on filesystem when not building x11rdp, improve text output

+++ X11rdp-o-matic.sh
@@ -46,6 +46,7 @@ TMPFILE=/tmp/xrdpver
 #XRDPGIT=https://github.com/ghomem/xrdp.git
 X11DIR=/opt/X11rdp
 WORKINGDIR=`pwd` # Would have used /tmp for this, but some distros I tried mount /tmp as tmpfs, and filled up.
+CONFIGUREFLAGS="--prefix=/usr --sysconfdir=/etc --localstatedir=/var --enable-fuse"
 
 if [ ! -e /usr/bin/dialog ]
 then
@@ -272,14 +273,14 @@ welcome_message()
 
 make_X11rdp_env()
 {
-MYDIR=$1
+X11DIR=$1
 WDIR=$2
-	  
+X11RDP=$3	  
 
-	if [ -e $MYDIR ]; then
-	  rm -rf $MYDIR
+	if [ -e $X11DIR ] && [ $X11RDP -eq 1 ]; then
+	  rm -rf $X11DIR
+	  mkdir -p $X11DIR
         fi
-	mkdir -p $MYDIR
 	
 	if [ -e $WDIR/xrdp ]; then
    	  rm -rf $WDIR/xrdp
@@ -327,7 +328,8 @@ rm -rf $BASEDIR/xrdp
 
 # checking latest xrdp version in master
 
-wget -O $TMPFILE $README
+echo
+wget -O $TMPFILE $README >& /dev/null
 VERSION=$(grep xrdp $TMPFILE | head -1 | cut -d " " -f2)
 rm -f $TMPFILE
 
@@ -337,8 +339,10 @@ echo " *** xrdp version is $VERSION ***"
 trap control_c SIGINT
 
 echo
-echo " *** Will remove the contents of $X11DIR and $WORKINGDIR/xrdp ***"
-echo
+if  [ "$X11RDP" == "1" ]; then
+    echo " *** Will remove the contents of $X11DIR and $WORKINGDIR/xrdp ***"
+    echo
+fi
 echo "Press ENTER to continue or CTRL+C to abort"
 read DUMMY
 clear
@@ -349,7 +353,7 @@ else
   INSTOPT="yes"
 fi
 
-make_X11rdp_env $X11DIR $WORKINGDIR
+make_X11rdp_env $X11DIR $WORKINGDIR $X11RDP
 
 calc_cpu_cores # find out how many cores we have to play with, and if >1, set a possible make command
 
@@ -370,7 +374,7 @@ then
 	  compile_X11rdp_interactive 
 	  package_X11rdp $VERSION $RELEASE $X11DIR
 	fi
-	compile_xrdp_interactive $VERSION $RELEASE $INSTOPT
+	compile_xrdp_interactive $VERSION $RELEASE $INSTOPT "$CONFIGUREFLAGS"
 else
 	download_xrdp_noninteractive $XRDPGIT
 	if [ "$PARALLELMAKE" == "1" ]
@@ -382,13 +386,15 @@ else
 	  compile_X11rdp_noninteractive 
 	  package_X11rdp $VERSION $RELEASE $X11DIR
 	fi
-	compile_xrdp_noninteractive $VERSION $RELEASE $INSTOPT
+	compile_xrdp_noninteractive $VERSION $RELEASE $INSTOPT "$CONFIGUREFLAGS"
 fi
 
 if [ "$INSTFLAG" == "0" ]; then
   # this is stupid but some Makefiles from X11rdp don't have an uninstall target (ex: Python!)
   # ... so instead of not installing X11rdp we remove it in the end
-  rm -rf $X11DIR
+  if  [ "$X11RDP" == "1" ]; then
+    rm -rf $X11DIR
+  fi
   if [ "$CLEANUP" == "1" ]; then
     cleanup $WORKINGDIR
   fi
